[AVR] Task 6: Complete.

+++ avr-2019/6/uart.h
@@ -1,6 +1,7 @@
 #pragma once
+#include <stdint.h>
 
-int USART_Receive();
+uint8_t USART_Receive();
 
 void USART_Init();
